+++ nullplatform/service_definition_agent_association/main.tf
@@ -10,25 +10,18 @@ resource "nullplatform_notification_channel" "channel_from_template" {
 
 
   configuration {
-    dynamic "agent" {
-      for_each = var.agent_command != null ? [1] : []
-      content {
-        api_key = var.api_key
-        command {
-          type = var.agent_command.type
-          data = {
-            cmdline = join(" ", compact([
-              var.agent_command.data.cmdline,
-              var.workflow_override_path != null ? "--overrides-path=${var.workflow_override_path}" : "",
-              var.service_path != null ? "--service-path=${var.service_path}" : "",
-            ]))
-            arguments   = jsonencode(try(var.agent_command.data.arguments, []))
-            environment = jsonencode(try(var.agent_command.data.environment, {}))
-          }
+    agent {
+      api_key = var.api_key
+      command {
+        type = "exec"
+        data = {
+          cmdline     = "${var.base_clone_path}/${var.repository_service_spec_repo}/${var.service_path}/entrypoint/entrypoint"
+          arguments   = jsonencode(var.agent_arguments)
+          environment = jsonencode({ NP_ACTION_CONTEXT = "'$${NOTIFICATION_CONTEXT}'" })
         }
-
-        selector = var.tags_selectors
       }
+
+      selector = var.tags_selectors
     }
   }