+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [1.47.0](https://github.com/nullplatform/tofu-modules/compare/v1.46.0...v1.47.0) (2026-03-26)
+
+
+### Features
+
+* **vpc:** export security group IDs as output ([#258](https://github.com/nullplatform/tofu-modules/issues/258)) ([7509399](https://github.com/nullplatform/tofu-modules/commit/750939944c9c13b2e4e6447587173e10a86d5f27))
+
+## [1.46.0](https://github.com/nullplatform/tofu-modules/compare/v1.45.0...v1.46.0) (2026-03-20)
+
+
+### Features
+
+* **nullplatform/cloud/aws/vpc:** implement aws-networking-configuration provider config ([#255](https://github.com/nullplatform/tofu-modules/issues/255)) ([3c3439b](https://github.com/nullplatform/tofu-modules/commit/3c3439b5ddde154636b696404169d0a0117e3262))
+
+## [1.45.0](https://github.com/nullplatform/tofu-modules/compare/v1.44.0...v1.45.0) (2026-03-19)
+
+
+### Features
+
+* **service:** add support to gitlab ([#249](https://github.com/nullplatform/tofu-modules/issues/249)) ([1d41de6](https://github.com/nullplatform/tofu-modules/commit/1d41de6c312008f82a77bc51b4d27962d9c825d0))
+
 ## [1.44.0](https://github.com/nullplatform/tofu-modules/compare/v1.43.0...v1.44.0) (2026-03-17)